requirements/requirements_develop.txt
asgiref==3.2.10
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.0
firebase-admin==4.3.0
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.1
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
httplib2==0.18.1
idna==2.10
msgpack==1.0.0
protobuf==3.12.2
psycopg2-binary==2.8.5
pudb==2019.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.25.9
urwid==2.1.0
